# openapi.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/products/{id}": {
      "get": {
        "operationId": "ProductsController_findOne",
        "summary": "Get product by ID",
        "description": "Retrieve a specific product by ID with all details including variations",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "description": "Product ID",
            "schema": {
              "type": "number"
            }
          },
          {
            "name": "includeVariations",
            "required": false,
            "in": "query",
            "description": "Include product variations in response",
            "schema": {
              "default": true,
              "type": "boolean"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Product found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProductResponseDto"
                }
              }
            }
          },
          "404": {
            "description": "Product not found"
          }
        },
        "tags": [
          "products"
        ],
        "security": [
          {
            "JWT-auth": []
          }
        ]
      },
      "patch": {
        "operationId": "ProductsController_update",
        "summary": "Update product",
        "description": "Update product information",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "description": "Product ID",
            "schema": {
              "type": "number"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateProductDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Product updated successfully",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProductResponseDto"
                }
              }
            }
          },
          "404": {
            "description": "Product not found"
          }
        },
        "tags": [
          "products"
        ],
        "security": [
          {
            "JWT-auth": []
          }
        ]
      },
      "delete": {
        "operationId": "ProductsController_remove",
        "summary": "Delete product",
        "description": "Delete a product",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "description": "Product ID",
            "schema": {
              "type": "number"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Product deleted successfully"
          },
          "404": {
            "description": "Product not found"
          },
          "409": {
            "description": "Cannot delete product with existing decants"
          }
        },
        "tags": [
          "products"
        ],
        "security": [
          {
            "JWT-auth": []
          }
        ]
      }
    },
...
    "/product-options": {
      "post": {
        "operationId": "ProductOptionsController_createOption",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateProductOptionDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": ""
          }
        },
        "tags": [
          "product-options"
        ],
        "security": [
          {
            "JWT-auth": []
          }
        ]
      },
      "get": {
        "operationId": "ProductOptionsController_findAllOptions",
        "parameters": [
          {
            "name": "productType",
            "required": true,
            "in": "query",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "product-options"
        ],
        "security": [
          {
            "JWT-auth": []
          }
        ]
      }
    },
    "/product-options/{id}": {
      "get": {
        "operationId": "ProductOptionsController_findOneOption",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "product-options"
        ],
        "security": [
          {
            "JWT-auth": []
          }
        ]
      },
      "patch": {
        "operationId": "ProductOptionsController_updateOption",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateProductOptionDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "product-options"
        ],
        "security": [
          {
            "JWT-auth": []
          }
        ]
      },
      "delete": {
        "operationId": "ProductOptionsController_removeOption",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "product-options"
        ],
        "security": [
          {
            "JWT-auth": []
          }
        ]
      }
    },
    "/product-options/values": {
      "post": {
        "operationId": "ProductOptionsController_createOptionValue",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateProductOptionValueDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": ""
          }
        },
        "tags": [
          "product-options"
        ],
        "security": [
          {
            "JWT-auth": []
          }
        ]
      }
    },
    "/product-options/values/all": {
      "get": {
        "operationId": "ProductOptionsController_findAllOptionValues",
        "parameters": [],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "product-options"
        ],
        "security": [
          {
            "JWT-auth": []
          }
        ]
      }
    },
    "/product-options/values/option/{optionId}": {
      "get": {
        "operationId": "ProductOptionsController_findOptionValuesByOption",
        "parameters": [
          {
            "name": "optionId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "product-options"
        ],
        "security": [
          {
            "JWT-auth": []
          }
        ]
      }
    },
    "/product-options/values/{id}": {
      "get": {
        "operationId": "ProductOptionsController_findOneOptionValue",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "product-options"
        ],
        "security": [
          {
            "JWT-auth": []
          }
        ]
      },
      "patch": {
        "operationId": "ProductOptionsController_updateOptionValue",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "product-options"
        ],
        "security": [
          {
            "JWT-auth": []
          }
        ]
      },
      "delete": {
        "operationId": "ProductOptionsController_removeOptionValue",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "product-options"
        ],
        "security": [
          {
            "JWT-auth": []
          }
        ]
      }
    },
...
  },
  "info": {
    "title": "E-commerce API",
    "description": "E-commerce API with NestJS - Products, Categories, Users, and Authentication",
    "version": "1.0",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "securitySchemes": {
      "JWT-auth": {
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "type": "http",
        "name": "JWT",
        "description": "Enter JWT token",
        "in": "header"
      }
    },
    "schemas": {
...
      "ProductResponseDto": {
        "type": "object",
        "properties": {}
      },
      "UpdateProductDto": {
        "type": "object",
        "properties": {}
      },
      "CreateProductOptionDto": {
        "type": "object",
        "properties": {}
      },
      "UpdateProductOptionDto": {
        "type": "object",
        "properties": {}
      },
      "CreateProductOptionValueDto": {
        "type": "object",
        "properties": {}
      },
...
    }
  }
}